Refactoring user components.
