+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
-import LoadingProvider from '@/src/loading/provider';
+import LoadingProvider from '@/src/loading/suspense';
 import HeaderProvider from '@/src/header.tsx/provider';
 import ErrorRecordProvider from '@/src/error/provider';
 import UidProvider from '@/src/user/provider';
