+++ src/error/handleError.tsx
@@ -1,0 +1,6 @@
+import { ErrorRecord, NextJsError } from './types';
+
+export default function handleError(
+  error: NextJsError,
+  errorStatus: ErrorRecord
+) {}
