+++ src/navigation/buttons/back.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+
+export default function BackButton() {
+  const router = useRouter();
+
+  return <button onClick={() => router.back()}>Back</button>;
+}
