+++ src/navigation/buttons/home.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+
+export default function HomeButton() {
+  const router = useRouter();
+
+  return <button onClick={() => router.push('/')}>Home</button>;
+}
