+++ src/navigation/buttons/refresh.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+
+export default function RefreshButton() {
+  const router = useRouter();
+
+  return <button onClick={() => router.refresh()}>Refresh</button>;
+}
