+++ src/user/context.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { createContext } from 'react';
 import { Uid } from './types';
